processing/histogram/requirements.txt
tabulate==0.8.6
tqdm==4.40.1
imageio==2.6.1
Pillow==7.0.0